Refactored tests. Better reuse
