+++ tests/test_engine_manager.py
@@ -10,153 +10,15 @@ from PyQt6.QtWidgets import QApplication
 from eloGraf.engine_manager import EngineManager
 from eloGraf.status import DictationStatus
 from eloGraf.stt_engine import STTController, STTProcessRunner
+from tests.helpers import (
+    AutoStoppingRunner,
+    DummyController,
+    DummyRunner,
+    DummySettings,
+    HangingRunner,
+)
 
 
-class DummySettings:
-    def __init__(self) -> None:
-        self.sttEngine = "dummy"
-
-    def get_engine_settings(self, engine_type: str):
-        from eloGraf.base_settings import EngineSettings
-        return EngineSettings(engine_type=engine_type, device_name="default")
-
-
-class DummyController(STTController):
-    def __init__(self) -> None:
-        self._exit_listeners: list = []
-
-    def add_state_listener(self, callback) -> None:
-        pass
-
-    def add_output_listener(self, callback) -> None:
-        pass
-
-    def add_exit_listener(self, callback) -> None:
-        self._exit_listeners.append(callback)
-
-    def remove_exit_listener(self, callback) -> None:
-        try:
-            self._exit_listeners.remove(callback)
-        except ValueError:
-            pass
-
-    def start(self) -> None:
-        pass
-
-    def stop_requested(self) -> None:
-        pass
-
-    def suspend_requested(self) -> None:
-        pass
-
-    def resume_requested(self) -> None:
-        pass
-
-    def fail_to_start(self) -> None:
-        pass
-
-    def handle_output(self, line: str) -> None:
-        pass
-
-    def handle_exit(self, return_code: int) -> None:
-        for listener in self._exit_listeners:
-            listener(return_code)
-
-    def transition_to(self, state: str) -> None:
-        pass
-
-    def emit_transcription(self, text: str) -> None:
-        pass
-
-    def emit_error(self, message: str) -> None:
-        pass
-
-    def get_status_string(self) -> str:
-        return "Dummy Controller"
-
-    @property
-    def dictation_status(self) -> DictationStatus:
-        return DictationStatus.IDLE
-
-
-class HangingRunner(STTProcessRunner):
-    def __init__(self) -> None:
-        self.running = True
-        self.stop_calls = 0
-        self.force_stop_calls = 0
-
-    def start(self, command: Sequence[str], env: Optional[Dict[str, str]] = None) -> bool:
-        self.running = True
-        return True
-
-    def stop(self) -> None:
-        self.stop_calls += 1
-
-    def force_stop(self) -> None:
-        self.force_stop_calls += 1
-        self.running = False
-
-    def suspend(self) -> None:
-        pass
-
-    def resume(self) -> None:
-        pass
-
-    def poll(self) -> None:
-        pass
-
-    def is_running(self) -> bool:
-        return self.running
-
-
-class AutoStoppingRunner(STTProcessRunner):
-    def __init__(self) -> None:
-        self.running = True
-        self.stop_calls = 0
-
-    def start(self, command: Sequence[str], env: Optional[Dict[str, str]] = None) -> bool:
-        self.running = True
-        return True
-
-    def stop(self) -> None:
-        self.stop_calls += 1
-        self.running = False
-
-    def suspend(self) -> None:
-        pass
-
-    def resume(self) -> None:
-        pass
-
-    def poll(self) -> None:
-        pass
-
-    def is_running(self) -> bool:
-        return self.running
-
-
-class DummyRunner(STTProcessRunner):
-    def __init__(self) -> None:
-        self.running = False
-
-    def start(self, command: Sequence[str], env: Optional[Dict[str, str]] = None) -> bool:
-        self.running = True
-        return True
-
-    def stop(self) -> None:
-        self.running = False
-
-    def suspend(self) -> None:
-        pass
-
-    def resume(self) -> None:
-        pass
-
-    def poll(self) -> None:
-        pass
-
-    def is_running(self) -> bool:
-        return self.running
 
 
 @pytest.fixture(scope="module")
